Hoàn thiện tính năng Chat: Sửa lỗi mất tin nhắn Shop, thêm bộ lọc từ ngữ thô tục, cải thiện giao diện các nút tương tác

+++ src/main/java/com/project/web_hand_made_TMDT/model/ChatPayload.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class ChatPayload {
     private String roomId;
-    private Long senderId;
+    private int senderId;
     private String senderName;
     private Integer senderRole;
     private String type; // "text" or "product"
@@ -13,9 +13,9 @@ public class ChatPayload {
     private String productInfo;
     
     // Room info
-    private Long customerId;
+    private int customerId;
     private String customerName;
     private String customerAvatar;
-    private Long shopId;
+    private int shopId;
     private String shopName;
 }
